fstab: Mount boot partition with default options.
Mounting them read-only was a good idea, but the package manager is
struggling with these as it figures out the space available before
attempting to install anything. For a read-only partition, it
determines that there is no space available and stops with an error.

Working around this may not be trivial and not be worth it, so we go
with writable boot partitions.