recovery: Remove editors, tools, and networking from image.
The cpio grew too big and U-Boot ran out of RAM when trying to
decompress the XZ-compressed image. It was just around 1 MiB in
compressed size, but that was enough to kill it. Fortunately, we
discovered this problem now, and not after roll-out in the field.

No networking also means no need for firmware BLOBs anymore, so the
image is pretty small now. However, no networking also means no logs
in the web browser. We will resort to serial logging during recovery,
which should be acceptable.