kernel: Switch to Linux 5.10.22.
SMP is *disabled* in this kernel configuration because our WLAN
adapter doesn't work very well in an SMP environment. Sadly, this
means that 3 out of 4 cores are not going to used on the Compute
Module 3+ because we cannot afford to maintain multiple kernels for
the time being.

While at it, exFAT file system has been enabled and two kernel
hardening options have been enabled.