recovery: Reduce xz compression level.
By default, Yocto Project uses the maximum compression level of 9, but
this requires the decompressor to consume more than 64 MiB of RAM just
for the decompression dictionary. That is too much for our CM1, and
thus the kernel states tells us that the XZ-compressed cpio image is
corrupt. It is not, but there is simply not enough RAM available for
decompression, leading to a misguiding error message.

We use compression level 1 now, which works well. The image size grows
by around 25%, but at least it works.