kernel: Restrict maximum rate to 192kHz.
The I2S driver supports 384kHz since kernel version 4.15, but we must
restrict to 192kHz for the time being because of Roon certification
(some of our appliances do not support 384kHz, and we are not allowed
to report what doesn't work).

The restriction to 192kHz is done in the CODEC driver (where it
belongs), not in the I2S driver.