gupnp-dlna: Bump PR because of some strange GStreamer dependency.
The package was rebuilt with a dependency on GStreamer >= 1.22, even
though nothing has changed. There shouldn't be a version number for
GStreamer coded into the RPM, and the package shouldn't have been
rebuilt, but because it was and was using the exact same name as
before, it broke our RPM repository.

As a quick fix, simply bump PR so that we can push out our release.